Add files via upload
TO DO LIST WEBSITE


🖥 Tech Stack Used:
- HTML: Structuring the web page.
- CSS: Styling for a visually appealing interface.
- JavaScript: Adding interactivity and functionality.


🎨 Design:
The design is user-friendly and intuitive, ensuring a seamless experience for anyone using the to-do list. Clean and minimalistic, the interface focuses on enhancing productivity without unnecessary distractions.

🌟 Key Features:
1. Add Tasks: Easily add new tasks with a click.
3. Remove Tasks: Delete completed or unwanted tasks effortlessly.

🔍 *User Experience:*
I prioritized a smooth user experience to make managing tasks a breeze. The to-do list adapts to different screen sizes, making it accessible from desktops, tablets, or smartphones.